refactoring: rename packet: .locator -> .object
