+++ src/main/java/org/bricolages/streaming/Application.java
@@ -2,7 +2,7 @@ package org.bricolages.streaming;
 import org.bricolages.streaming.filter.*;
 import org.bricolages.streaming.event.*;
 import org.bricolages.streaming.stream.*;
-import org.bricolages.streaming.locator.*;
+import org.bricolages.streaming.object.*;
 import org.bricolages.streaming.exception.*;
 import org.bricolages.streaming.preflight.ReferenceGenerator;
 import org.bricolages.streaming.preflight.Runner;
