+++ src/main/java/org/bricolages/streaming/Preprocessor.java
@@ -1,7 +1,7 @@
 package org.bricolages.streaming;
 import org.bricolages.streaming.event.*;
 import org.bricolages.streaming.stream.*;
-import org.bricolages.streaming.locator.*;
+import org.bricolages.streaming.object.*;
 import org.bricolages.streaming.exception.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import java.io.BufferedReader;
