+++ src/main/java/org/bricolages/streaming/object/LocatorParseException.java
@@ -1,4 +1,4 @@
-package org.bricolages.streaming.locator;
+package org.bricolages.streaming.object;
 import org.bricolages.streaming.exception.ApplicationException;
 
 public class LocatorParseException extends ApplicationException {
