+++ src/main/java/org/bricolages/streaming/preflight/Runner.java
@@ -3,8 +3,8 @@ import org.bricolages.streaming.preflight.definition.*;
 import org.bricolages.streaming.stream.PacketFilterFactory;
 import org.bricolages.streaming.stream.PacketFilter;
 import org.bricolages.streaming.stream.PacketFilterLog;
-import org.bricolages.streaming.locator.S3ObjectLocator;
-import org.bricolages.streaming.locator.LocatorIOException;
+import org.bricolages.streaming.object.S3ObjectLocator;
+import org.bricolages.streaming.object.LocatorIOException;
 import org.bricolages.streaming.exception.*;
 import java.io.FileReader;
 import java.io.IOException;
