+++ src/main/java/org/bricolages/streaming/stream/PacketFilter.java
@@ -3,7 +3,10 @@ import org.bricolages.streaming.stream.processor.StreamColumnProcessor;
 import org.bricolages.streaming.filter.Op;
 import org.bricolages.streaming.filter.Record;
 import org.bricolages.streaming.filter.JSONException;
-import org.bricolages.streaming.locator.*;
+import org.bricolages.streaming.object.LocatorIOManager;
+import org.bricolages.streaming.object.S3ObjectLocator;
+import org.bricolages.streaming.object.S3ObjectMetadata;
+import org.bricolages.streaming.object.LocatorIOException;
 import java.util.List;
 import java.io.IOException;
 import java.io.UncheckedIOException;
